requirements.txt
# AlphaAI - AI-Native Hedge Fund Platform
# Core Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Market Data
yfinance==0.2.51
pandas==2.2.3
pandas-ta==0.3.14b1
numpy==1.26.4

# News & Sentiment
finnhub-python==2.4.20
vaderSentiment==3.3.2
transformers==4.47.1
torch==2.5.1
feedparser==6.0.11

# LLM / AI Agent Orchestration
langchain==0.3.14
langchain-openai==0.3.0
langchain-google-genai==2.0.8
langchain-groq==0.2.4

# Database
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Report Generation
jinja2==3.1.5
markdown==3.7.1
